requirements.txt
Flask==1.0.2
Flask-Migrate==2.4.0
Flask-SQLAlchemy==2.4.0
gunicorn==19.9.0
psycopg2-binary==2.8.2
python-dotenv==0.10.3
slackclient==2.1.0
cachetools==3.1.1
requests==2.22.0